We now have the User and Group databases displaying in the relevant locations, and the database is now being seeded with test data if the Database Model changes or does not the database doesn't exist.
Despite having the require schema there are still other tables to add, for example a branch table in order to handle the many to many relationship between Users and Groups. Also, to keep the password table consistent with the scheme a separate table for salts will be necessary. This Salt table will have a one-to-one relationship with the Users table.

Note how as of right now the password is inserted into the database without any hashing done. I plan on using Argon2 to hash the passwords and the salt will be provided by RNGCryptoServiceProvider.

Argon2 had won the 2015 Password Hashing Competition where a Computer Scientist from The University of Auckland, where I studied, was a judge.

Also, many other legacy hashing technologies such as MD5, SHA1, SHA256, SHA512, etc, are fast but considered outdated and hence not as great for password storage.